+++ lib/src/visitor/interface/expression.dart
@@ -12,6 +12,7 @@ import '../../ast/sass.dart';
 abstract class ExpressionVisitor<T> {
   T visitBinaryOperationExpression(BinaryOperationExpression node);
   T visitBooleanExpression(BooleanExpression node);
+  T visitCalculationExpression(CalculationExpression node);
   T visitColorExpression(ColorExpression node);
   T visitInterpolatedFunctionExpression(InterpolatedFunctionExpression node);
   T visitFunctionExpression(FunctionExpression node);